Consolidate formulaToExpression parts
